Replacing JsonLoad with the Repository class
